DigitNavigation -> DigitHeader, if there is no renderDrawer, the drawer is removed
