+++ components/elements/digit-header/index.js
@@ -1,0 +1,2 @@
+import DigitHeader from "./DigitHeader.element";
+export default DigitHeader;
