--- components/elements/digit-navigation/index.js
@@ -1,2 +1,0 @@
-import DigitNavigation from "./DigitNavigation.element";
-export default DigitNavigation;
