+++ components/index.js
@@ -9,7 +9,7 @@ import DigitFormField from "./elements/digit-form-field";
 import DigitFormFieldArray from "./elements/digit-form-field-array";
 import DigitTextArea from "./elements/digit-text-area";
 import DigitSelect from "./elements/digit-select";
-import DigitNavigation from "./elements/digit-navigation";
+import DigitHeader from "./elements/digit-header";
 import DigitNavLink from "./elements/digit-nav-link";
 import DigitFAB from "./elements/digit-fab";
 import DigitDisplayData from "./elements/digit-display-data";
@@ -46,7 +46,7 @@ export {
   DigitFormFieldArray,
   DigitTextArea,
   DigitSelect,
-  DigitNavigation,
+  DigitHeader,
   DigitNavLink,
   DigitFAB,
   DigitDisplayData,
